MainFragment: Add option to enable auto-checking of server OR disable auto-checking
Since with my development criteria, UI Constants should be its UI Class, but this autochecking needs to be shared between
MainPreferenceFragment and AdvancedServerSettings. So share them.

At least we are not sharing dependency[UI Constants is val-String] thus even if AdvancedServerSettings are NOT called, meaning
mServerAutoChecking is null, sharedpreference settings might not be null.[if null or not set, it will automatically set to false though]